mv bert fuse classes to vlfuse helper
